+++ optionsScreen.go
@@ -1,13 +1,14 @@
 package hashcatlauncher
 
 import (
+	"fmt"
 	"strconv"
 	"fyne.io/fyne"
 	"fyne.io/fyne/widget"
 	dialog2 "github.com/sqweek/dialog"
 )
 
-func optionsScreen(app fyne.App, hcl_gui *hcl_gui) fyne.CanvasObject {
+func optionsScreen(app fyne.App, hcl_gui *hcl_gui, hash_type_fakeselector *widget.Box) fyne.CanvasObject {
 	hcl_gui.hc_binary_file_select = widget.NewSelect([]string{"Browse..."}, func(string){
 		file, err := dialog2.File().Title("Select hashcat binary file").Filter("Bin/Exe Files", "exe", "bin").Load()
 		if err == nil {
@@ -46,8 +47,6 @@ func optionsScreen(app fyne.App, hcl_gui *hcl_gui) fyne.CanvasObject {
 	hcl_gui.hc_extra_args = widget.NewEntry()
 	hcl_gui.hc_extra_args.SetPlaceHolder("(Advanced Only)")
 	hcl_gui.hc_extra_args.OnChanged = func(s string) {
-		s = s[:min(len(s), 40)]
-		hcl_gui.hc_extra_args.SetText(s)
 		SetPreference_hashcat_extra_args(app, hcl_gui, s)
 	}
 
@@ -56,13 +55,39 @@ func optionsScreen(app fyne.App, hcl_gui *hcl_gui) fyne.CanvasObject {
 			widget.NewForm(
 				widget.NewFormItem("Binary:", hcl_gui.hc_binary_file_select),
 				widget.NewFormItem("Status Timer:", hcl_gui.hc_status_timer_select),
-				widget.NewFormItem("Extra Args:", hcl_gui.hc_extra_args),
+				widget.NewFormItem("Extra Args:", widget.NewHScrollContainer(hcl_gui.hc_extra_args)),
 			),
 		),
 		widget.NewGroup("hashcat.launcher options",
 			widget.NewForm(
 				widget.NewFormItem("Auto Session Start:", hcl_gui.autostart_sessions_select),
 				widget.NewFormItem("Max Active Sessions:", hcl_gui.max_active_sessions_select),
+			),
+		),
+		widget.NewGroup("hashcat.launcher appearance",
+			widget.NewForm(
+				widget.NewFormItem("Theme:", 
+					widget.NewSelect([]string{"Light", "Dark"}, func(theme string) {
+						switch theme {
+						case "Light":
+							hcl_gui.Settings.SetTheme("light")
+						case "Dark":
+							hcl_gui.Settings.SetTheme("dark")
+						}
+
+						go get_available_hash_typess(hcl_gui)
+						if hcl_gui.hashcat.args.hash_type >= 0 {
+							fake_hash_type_selector_hack(hcl_gui, hash_type_fakeselector, fmt.Sprintf("%d", hcl_gui.hashcat.args.hash_type))
+						} else {
+							fake_hash_type_selector_hack(hcl_gui, hash_type_fakeselector, "(Select one)")
+						}
+					}),
+				),
+				widget.NewFormItem("Scaling:",
+					widget.NewSelect([]string{"auto", "50%", "75%", "80%", "90%", "100%", "110%", "125%", "150%", "175%", "200%"}, func(value string) {
+						hcl_gui.Settings.SetScale(value)
+					}),
+				),
 			),
 		),
 		widget.NewGroup("Notes",