+++ docs/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project (hashcat.launcher) will be documented in this file.
 
+## [0.1.2] - 25/04/2020
+### Fixed:
+ - Spelling errors
+ - Fields were limited and can't type long arguments
+
+### Changed:
+ - Hash Type search field is now fixed at the top
+ - Merged appearanceScreen to optionsScreen
+
+### Added:
+ - Mask attack supports hcmask files
+ - Mask length counter
+ - Help messages
+
 ## [0.1.1] - 21/04/2020
 ### Fixed:
  - Tabs texts were not being updated